requirements.txt
appnope==0.1.0
backcall==0.1.0
Click==7.0
decorator==4.4.1
Flask==1.1.1
gunicorn==20.0.4
ipdb==0.12.3
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
mock==3.0.5
nltk==3.4.5
numpy==1.18.1
pandas==0.24.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.4.1
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
traitlets==4.3.3
wcwidth==0.1.8
Werkzeug==1.0.0